requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
networkx==3.2.1
geopandas==0.13.2
fiona==1.9.4
shapely==2.0.2
pydantic==2.5.0
python-multipart==0.0.6 
scipy==1.16.1
numpy<2 